feat(app): import projects from Github
